requirements.txt
Django
djangorestframework
djangorestframework-jwt
drf-nested-routers
python-dateutil
django-cors-headers
